refactor: channel service
